only query realms endpoint once
currently the realm list query is expensive and we only need it once.
