+++ src/PageNav.tsx
@@ -12,19 +12,13 @@ import {
 
 import { RealmSelector } from "./components/realm-selector/RealmSelector";
 import { useRealm } from "./context/realm-context/RealmContext";
-import { DataLoader } from "./components/data-loader/DataLoader";
-import { useAdminClient } from "./context/auth/AdminClient";
 import { useAccess } from "./context/access/Access";
 import { routes } from "./route-config";
 
 export const PageNav: React.FunctionComponent = () => {
   const { t } = useTranslation("common");
   const { hasAccess, hasSomeAccess } = useAccess();
-  const { realm } = useRealm();
-  const adminClient = useAdminClient();
-  const realmLoader = async () => {
-    return _.sortBy(await adminClient.realms.find(), "realm");
-  };
+  const { realm, realms } = useRealm();
 
   const history = useHistory();
 
@@ -83,13 +77,9 @@ export const PageNav: React.FunctionComponent = () => {
       nav={
         <Nav onSelect={onSelect}>
           <NavList>
-            <DataLoader loader={realmLoader} deps={[realm]}>
-              {(realmList) => (
-                <NavItem className="keycloak__page_nav__nav_item__realm-selector">
-                  <RealmSelector realmList={realmList || []} />
-                </NavItem>
-              )}
-            </DataLoader>
+            <NavItem className="keycloak__page_nav__nav_item__realm-selector">
+              <RealmSelector realmList={realms} />
+            </NavItem>
           </NavList>
           {isOnAddRealm() && (
             <NavGroup title="">
