+++ src/context/realm-context/RealmContext.tsx
@@ -1,14 +1,19 @@
-import React, { useContext, useState } from "react";
+import RealmRepresentation from "keycloak-admin/lib/defs/realmRepresentation";
+import React, { useContext, useEffect, useState } from "react";
 import { RecentUsed } from "../../components/realm-selector/recent-used";
+import { useErrorHandler } from "react-error-boundary";
+import { asyncStateFetch, useAdminClient } from "../auth/AdminClient";
 
 type RealmContextType = {
   realm: string;
   setRealm: (realm: string) => void;
+  realms: RealmRepresentation[];
 };
 
 export const RealmContext = React.createContext<RealmContextType>({
   realm: "",
   setRealm: () => {},
+  realms: [],
 });
 
 type RealmContextProviderProps = { children: React.ReactNode };
@@ -17,6 +22,9 @@ export const RealmContextProvider = ({
   children,
 }: RealmContextProviderProps) => {
   const [realm, setRealm] = useState("");
+  const [realms, setRealms] = useState<RealmRepresentation[]>([]);
+  const adminClient = useAdminClient();
+  const errorHandler = useErrorHandler();
   const recentUsed = new RecentUsed();
 
   const set = (realm: string) => {
@@ -24,8 +32,29 @@ export const RealmContextProvider = ({
     setRealm(realm);
   };
 
+  useEffect(() => {
+    return asyncStateFetch(
+      () => adminClient.realms.find(),
+      (realms) => {
+        setRealms(realms);
+      },
+      errorHandler
+    );
+  }, []);
+
+  const set = (realm: string) => {
+    if (
+      realms.length === 0 ||
+      realms.findIndex((r) => r.realm == realm) !== -1
+    ) {
+      setRealm(realm);
+      adminClient.setConfig({
+        realmName: realm,
+      });
+    }
+  };
   return (
-    <RealmContext.Provider value={{ realm, setRealm: set }}>
+    <RealmContext.Provider value={{ realm, setRealm: set, realms }}>
       {children}
     </RealmContext.Provider>
   );
